+++ app/src/main/java/os/com/ui/contest/adapter/AllContestAdapter.kt
@@ -54,7 +54,7 @@ class AllContestAdapter(
 
         if (!contest.get(holder.adapterPosition).total_teams.isEmpty() && !contest.get(holder.adapterPosition).teams_joined.isEmpty()) {
             val strtValue =
-                contest.get(holder.adapterPosition).total_teams.toInt() - contest.get(holder.adapterPosition).teams_joined.toInt()
+                contest.get(holder.adapterPosition).total_teams.toLong() - contest.get(holder.adapterPosition).teams_joined.toInt()
             holder.itemView.txt_StartValue.text = mContext.getString(R.string.only) + " " + strtValue.toString() + " " +
                     mContext.getString(R.string.spots_left)
             holder.itemView.crs_Progress.setMinValue(0f)
@@ -66,7 +66,7 @@ class AllContestAdapter(
         if (!contest.get(holder.adapterPosition).multiple_team!!) {
             if (contest.get(holder.adapterPosition).is_joined!!) {
                 val total_teams =
-                    contest.get(holder.adapterPosition).total_teams.toInt() - contest.get(holder.adapterPosition).teams_joined.toInt()
+                    contest.get(holder.adapterPosition).total_teams.toLong() - contest.get(holder.adapterPosition).teams_joined.toInt()
                 if (total_teams > 0) {
                     holder.itemView.txt_Join.text = mContext.getString(R.string.invite)
                 } else {
@@ -79,7 +79,7 @@ class AllContestAdapter(
             if (contest.get(holder.adapterPosition).is_joined!!) {
                 holder.itemView.txt_Join.text = mContext.getString(R.string.join_plus)
                 val total_teams =
-                    contest.get(holder.adapterPosition).total_teams.toInt() - contest.get(holder.adapterPosition).teams_joined.toInt()
+                    contest.get(holder.adapterPosition).total_teams.toLong() - contest.get(holder.adapterPosition).teams_joined.toInt()
                 if (total_teams > 0) {
                     holder.itemView.txt_Join.text = mContext.getString(R.string.join_new)
                 } else {
